requirements.txt
requests
jsonschema
curlify
pytest==7.1.2
allure-pytest==2.9.45
python-dotenv==0.20.0